pyproject.toml
[project]
name = "mimi-smollm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.0",
    "bitsandbytes>=0.45.5",
    "datasets[audio]>=4.0.0",
    "flash-attn>=2.8.2",
    "librosa>=0.11.0",
    "moshi",
    "python-dotenv>=1.1.1",
    "soundfile>=0.13.1",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchcodec==0.5",
    "transformers~=4.55",
    "wandb>=0.21.1",
]


[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true


[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchcodec = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
moshi = { git = "https://github.com/jspaulsen/moshi", subdirectory = "moshi" }
